% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adamReg.R
\name{cross_validate_gd}
\alias{cross_validate_gd}
\title{Cross-validation of gradient descent for optimal lambda}
\usage{
cross_validate_gd(
  k = 5,
  X,
  Y,
  alpha = 0.001,
  penalty,
  maxit = 1000,
  tol = 1e-06
)
}
\arguments{
\item{k}{Number of folds used.}

\item{X}{A matrix of features.}

\item{Y}{The response vector, binary.}

\item{alpha}{The learning rate (step size), fixed.}

\item{penalty}{The penalty (L1, L2) for the loss function.}

\item{maxit}{The maximum number of iterations.}

\item{tol}{The tolerance for convergence.}
}
\value{
An optimal lambda value alongside a plot of all lambda values explored.
}
\description{
Cross-validation to identify the optimal lambda to be used in gradient descent optimization for estimating
the coefficients in logistic regression. See \link{logistic_regression_gd}.
}
