% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adamReg.R
\name{gradient_function}
\alias{gradient_function}
\title{Calculate gradient of the loss function for logistic regression.}
\usage{
gradient_function(theta, X, Y)
}
\arguments{
\item{theta}{A parameter vector.}

\item{X}{A matrix of features.}

\item{Y}{The response vector.}
}
\value{
A numeric value.
}
\description{
Calculates the gradient of the objective function for logistic regression.
}
\details{
The gradient without is: \deqn{-X^T(Y-\mu)}


where \eqn{\mu = \frac{e^{X^T\beta}}{1+e^{X^T\beta}}}.
}
